+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Please note that this is not a complete list of every change. It will be continu
 
 Changes
 -----------------------
+**--01/13/2026--**
+
+* server:<br />fix clientCommand spam when entering the world
+
 **--01/12/2026--**
 
 * server:<br />server snapshot cleanup